Build manticore from scratch, multi-arch build prep
Followed a few pointers of a guy that built manticore on alpine, it works and seems to properly compile.
This is mainly so I can cross-compile for different architectures at a later date. Since the official container only runs on x86_64.

Also, strangely when mem_limit is set to 512M, I can run indexer inside the container.
But if it's at anything or >~750MB indexer seems to die in the container with error 137 aka. OOM